CommonVoice24.0_asr/requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0
lazy_loader>=0.3
setuptools>=65.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
transformers>=4.30.0
soundfile>=0.12.0
requests>=2.31.0
python-dotenv>=1.0.0
pydub>=0.25.1

